Moving things from the screen to the game class so multiple screens work
